+++ example/lib/models/base_response.g.dart
@@ -9,21 +9,19 @@ part of 'base_response.dart';
 BaseResponse<T> _$BaseResponseFromJson<T>(
   Map<String, dynamic> json,
   T Function(Object? json) fromJsonT,
-) =>
-    BaseResponse<T>(
-      success: json['success'] as bool,
-      message: json['message'] as String,
-      data: DataConverter<T?>().fromJson(json['data']),
-      code: (json['code'] as num?)?.toInt(),
-    );
+) => BaseResponse<T>(
+  success: json['success'] as bool,
+  message: json['message'] as String,
+  data: DataConverter<T?>().fromJson(json['data']),
+  code: (json['code'] as num?)?.toInt(),
+);
 
 Map<String, dynamic> _$BaseResponseToJson<T>(
   BaseResponse<T> instance,
   Object? Function(T value) toJsonT,
-) =>
-    <String, dynamic>{
-      'success': instance.success,
-      'message': instance.message,
-      'data': DataConverter<T?>().toJson(instance.data),
-      'code': instance.code,
-    };
+) => <String, dynamic>{
+  'success': instance.success,
+  'message': instance.message,
+  'data': DataConverter<T?>().toJson(instance.data),
+  'code': instance.code,
+};