+++ example/lib/models/user.g.dart
@@ -6,12 +6,10 @@ part of 'user.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-User _$UserFromJson(Map<String, dynamic> json) => User(
-      id: (json['id'] as num).toInt(),
-      name: json['name'] as String,
-    );
+User _$UserFromJson(Map<String, dynamic> json) =>
+    User(id: (json['id'] as num).toInt(), name: json['name'] as String);
 
 Map<String, dynamic> _$UserToJson(User instance) => <String, dynamic>{
-      'id': instance.id,
-      'name': instance.name,
-    };
+  'id': instance.id,
+  'name': instance.name,
+};